Updating bphome.sh (backup making script)
It now takes list of directories that contains directories with git repositories.
For each such directory checks remote.origin.url field and if so excludes given
directory from backup and saves name of the directory and git remote url to file
in /home//repos

In the future there probably will be scirpt that will clone all those repos on
machine which uses given backup. Its usefull because some repos like kernels or
gccs or llvms contains massive bulk of object files which are not really needed
when doing backup.